Finished frontend

+++ .docker/entrypoint.sh
@@ -1,8 +1,24 @@
 #!/bin/bash
-
 # On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+
+### Front-end
+npm config set cache /var/www/.npm-cache --global
+cd frontend && npm install && cd ..
+
+### Back-end
+cd backend
+
+if [ ! -f ".env" ]; then
+  cp .env.example .env
+fi
+
+if [ ! -f ".env.testing" ]; then
+  cp .env.testing.example .env.testing
+fi
+
+if [ ! -f ".env.dusk.testing" ]; then
+  cp .env.dusk.testing.example .env.dusk.testing
+fi
 
 composer install
